refactor file structure for sidebar
