+++ src/components/index.ts
@@ -1,14 +1,9 @@
 // import Header from "./Header";
-import HideSideBarBtn from "./HideSideBarBtn";
+import HideSideBarBtn from "./sidebar/HideSideBarBtn";
 import Layout from "./Layout";
-import SwitchBtn from "./SwitchBtn";
-import ThemeSwitch from "./ThemeSwitch";
 
 
 export {
     Layout,
-    ThemeSwitch,
-    SwitchBtn,
-    HideSideBarBtn,
     // Header,
 }
